fix: class not found exception 해결
MessageTest 클래스의 이름을 PrintMessageTest로 만들어서 생겼던 에러
